Add and apply precommit hooks
