+++ .github/workflows/industrial_ci.yaml
@@ -1,22 +1,21 @@
+---
 name: Industrial CI
 on:
-  workflow_call:
-  push:
+    workflow_call:
+    push:
 
 jobs:
-  industrial_ci:
-    strategy:
-      fail-fast: false
-      matrix:
-        ROS_DISTRO: [humble]
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v3
-      - name: Clone installation requirements
-        shell: bash
-        run : python3 -m pip install -U vcstool &&
-              vcs import . < ./rosbot/rosbot_hardware.repos &&
-              vcs import . < ./rosbot/rosbot_simulation.repos
-      - uses: ros-industrial/industrial_ci@master
-        env:
-          ROS_DISTRO: ${{matrix.ROS_DISTRO}}
+    industrial_ci:
+        strategy:
+            fail-fast: false
+            matrix:
+                ROS_DISTRO: [humble]
+        runs-on: ubuntu-latest
+        steps:
+            - uses: actions/checkout@v3
+            - name: Clone installation requirements
+              shell: bash
+              run: python3 -m pip install -U vcstool && vcs import . < ./rosbot/rosbot_hardware.repos && vcs import . < ./rosbot/rosbot_simulation.repos
+            - uses: ros-industrial/industrial_ci@master
+              env:
+                  ROS_DISTRO: ${{matrix.ROS_DISTRO}}
