+++ .github/workflows/spellcheck.yaml
@@ -1,14 +1,15 @@
+---
 name: Spellcheck Action
 
 on:
-  workflow_call:
-  push:
+    workflow_call:
+    push:
 
 jobs:
-  build:
-    name: Spellcheck
-    runs-on: ubuntu-latest
-    steps:
-    - uses: actions/checkout@v3
-    - uses: rojopolis/spellcheck-github-actions@0.33.1
-      name: Spellcheck
+    build:
+        name: Spellcheck
+        runs-on: ubuntu-latest
+        steps:
+            - uses: actions/checkout@v3
+            - uses: rojopolis/spellcheck-github-actions@0.33.1
+              name: Spellcheck
