+++ .pre-commit-config.yaml
@@ -1,0 +1,70 @@
+---
+repos:
+    - repo: https://github.com/pre-commit/pre-commit-hooks
+      rev: v4.5.0
+      hooks:
+          - id: check-merge-conflict
+          - id: trailing-whitespace
+          - id: end-of-file-fixer
+          - id: check-yaml
+          - id: check-xml
+          - id: check-added-large-files
+          - id: check-ast
+          - id: check-json
+          - id: name-tests-test
+            files: ^.*\/test\/.*$
+            args: [--pytest-test-first]
+
+    - repo: https://github.com/codespell-project/codespell
+      rev: v1.16.0
+      hooks:
+          - id: codespell
+            name: codespell
+            description: Checks for common misspellings in text files.
+            entry: codespell  *
+            language: python
+            types: [text]
+
+    - repo: https://github.com/jumanjihouse/pre-commit-hook-yamlfmt
+      rev: 0.2.1
+      hooks:
+          - id: yamlfmt
+            files: ^.github|./\.yaml
+
+    - repo: https://github.com/psf/black
+      rev: 22.12.0
+      hooks:
+        - id: black
+          args: ["--line-length=99"]
+
+    - repo: https://github.com/PyCQA/flake8
+      rev: 6.1.0
+      hooks:
+        - id: flake8
+          args: ["--ignore=E501"] # ignore too long line, black checks it
+
+    - repo: local
+      hooks:
+      - id: ament_lint_cmake
+        name: ament_lint_cmake
+        description: Check format of CMakeLists.txt files.
+        entry: ament_lint_cmake
+        language: system
+        files: CMakeLists\.txt$
+
+    - repo: local
+      hooks:
+      - id: ament_copyright
+        name: ament_copyright
+        description: Check if copyright notice is available in all files.
+        stages: [commit]
+        entry: ament_copyright
+        language: system
+
+      # Docs - RestructuredText hooks
+    - repo: https://github.com/PyCQA/doc8
+      rev: v1.1.1
+      hooks:
+        - id: doc8
+          args: ['--max-line-length=100', '--ignore=D001']
+          exclude: ^.*\/CHANGELOG\.rst/.*$
