+++ rosbot/rosbot_simulation.repos
@@ -3,7 +3,7 @@ repositories:
     type: git
     url: https://github.com/ros-controls/gz_ros2_control.git
     # on branch humble hardware isn't activated
-    # recently on master API breaking change was introduced, it is necessay to use commit before this change
+    # recently on master API breaking change was introduced, it is necessary to use commit before this change
     version: b296ff2f5c3758b637a70bd496fe6ed875ab03ce
 
   husarion/husarion_office_gz:
