+++ rosbot_controller/launch/controller.launch.py
@@ -35,8 +35,7 @@ def generate_launch_description():
         "mecanum",
         default_value="False",
         description=(
-            "Whether to use mecanum drive controller "
-            "(otherwise diff drive controller is used)"
+            "Whether to use mecanum drive controller " "(otherwise diff drive controller is used)"
         ),
     )
 
@@ -154,12 +153,10 @@ def generate_launch_description():
     )
 
     # Delay start of robot_controller after joint_state_broadcaster
-    delay_robot_controller_spawner_after_joint_state_broadcaster_spawner = (
-        RegisterEventHandler(
-            event_handler=OnProcessExit(
-                target_action=joint_state_broadcaster_spawner,
-                on_exit=[robot_controller_spawner],
-            )
+    delay_robot_controller_spawner_after_joint_state_broadcaster_spawner = RegisterEventHandler(
+        event_handler=OnProcessExit(
+            target_action=joint_state_broadcaster_spawner,
+            on_exit=[robot_controller_spawner],
         )
     )
