+++ rosbot_controller/test/test_flake8.py
@@ -20,6 +20,4 @@ import pytest
 @pytest.mark.linter
 def test_flake8():
     rc, errors = main_with_errors(argv=[])
-    assert rc == 0, "Found %d code style errors / warnings:\n" % len(
-        errors
-    ) + "\n".join(errors)
+    assert rc == 0, "Found %d code style errors / warnings:\n" % len(errors) + "\n".join(errors)
