+++ rosbot_gazebo/launch/simulation.launch.py
@@ -35,8 +35,7 @@ def generate_launch_description():
         "mecanum",
         default_value="False",
         description=(
-            "Whether to use mecanum drive controller "
-            "(otherwise diff drive controller is used)"
+            "Whether to use mecanum drive controller " "(otherwise diff drive controller is used)"
         ),
     )
